Move tool-call-decision to its own file to prevent circular dependency
